Add code inline
This patch highlights the problem with the current text matching.   Text
matching needs to be a lot more restrictive, but it’s not obvious how
that should be achieved at the moment.

Another open question is what inlines should allow other inlines nested
within them.  Code shouldn’t.  Should emphasis?  Is it important to be
able to emphasize that /‹this is a code block›/?