Improve emphasis handling
We only check for ending '/' if we actually have to, that is, an
emphatic sequence has already been started.  This makes things faster
and, more importantly, much more correct in how we want it interpreted.
The cost is a more complex grammar, but seeing as how we have so few
inline rules, the added “duplication” doesn’t hurt that much.